Move `get_asset_path()` in a separate module in the `utils` folder
